macro_seq_v_par/scripts/requirements.txt
execo>=2.6.8
pymongo>=4.2.0
PyYAML>=5.4.1
python-dateutil
